% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_x.R
\name{cont_x}
\alias{cont_x}
\title{cont_x}
\usage{
cont_x(x)
}
\arguments{
\item{x}{vector of coordinates}
}
\value{
Returns a vector of velocities
}
\description{
\code{cont_x} create a continuous seq between the maximum and minimum values of a given vector. This is
useful for analysing how much an observed vector deviates from a continuous sequence.
}
\examples{


x <- sort(runif(5, 0, 10))
cont_x(x)

}
