% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eye_dat_vx}
\alias{eye_dat_vx}
\title{Raw eye-tracking data with x velocity column.}
\format{A data frame with 17750 rows and 4 variables:
\describe{
\item{RespID}{identifier for trial}
\item{ms}{time in ms}
\item{x,y}{coordinates in pixels}
\item{vx}{velocity from x position to next x position}
\item{new_line}{new line identifier}
\item{line_num}{line number identifier}
}}
\usage{
eye_dat_vx
}
\description{
A dataset containing raw eye-tracking data from a sequential reading task (\code{eye_data_clean})
prepared with a line number estimate.
}
\keyword{datasets}
