% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eye_data_noisy}
\alias{eye_data_noisy}
\title{Raw eye-tracking data.}
\format{A data frame with 17750 rows and 4 variables:
\describe{
\item{RespID}{identifier for trial}
\item{ms}{time in ms}
\item{x}{x coordinate}
\item{y}{y coordinate}
}}
\usage{
eye_data_noisy
}
\description{
A dataset containing raw eye-tracking data from a sequential reading task.
}
\keyword{datasets}
