% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_points.R
\name{filter_points}
\alias{filter_points}
\title{filter_points}
\usage{
filter_points(df, xmin, xmax, ymin, ymax, limits = NULL, x = x,
  y = y, margin = 25)
}
\arguments{
\item{df}{A data frame containing x and y coordinates}

\item{xmin, xmax, ymin, ymax}{Minimum and maximum coordinates of target area}

\item{limits}{a named list of limits including xmin, xmax, ymin, ymax}

\item{x, y}{x and y variables to use}

\item{margin}{The margin (in pixels) around the target area you want to include}
}
\value{
Returns a data frame with coordinates outside the target area replaced by NaN values
}
\description{
\code{filter_points} filters points outside the target area.
}
\examples{


df <- data.frame(x = runif(10, 0, 10), y = runif(10, 0, 10))
filter_points(df, 4, 6, 5, 7, margin = 1)

}
