% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_velocity.R
\name{get_velocity}
\alias{get_velocity}
\title{get_velocity}
\usage{
get_velocity(x, smooth = 1, method = 1)
}
\arguments{
\item{x}{vector of coordinates to calculate velocity between}

\item{smooth}{number of leading points to smooth velocity by. If smooth = 1 the output will not be smoothed}

\item{method}{the method to use. 1 is the velocity to the next point, 2 is based on both previous and next point but ignoring current point.}
}
\value{
Returns a vector of velocities
}
\description{
\code{get_velocity} calculate the velocity of a vector using either the velocity from one point to the next or the method used in the
saccades package which calculates velocity based on surrounding points.
}
\examples{


x <- sort(runif(5, 0, 10))
x <- c(x, rev(x))
get_velocity(x)

}
