% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_diff.R
\name{phase_diff}
\alias{phase_diff}
\title{phase_diff}
\usage{
phase_diff(x1, x2, lowlim = 0, highlim = 360)
}
\arguments{
\item{x1, x2}{x coordinates to calculate phase difference between}

\item{lowlim, highlim}{maximum and minimum expected x values. Defaults apply to degrees in a circle}
}
\value{
Returns a vector of values (between -0.5 and 0.5) indicating the phase difference between two x coordinates
}
\description{
\code{phase_diff} calcultes the phase difference between two x coordinates
}
\examples{


x1 <- seq(0, 360, 360/4)
x2 <- 180

phase_diff(x1, x2)

}
