% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_diff_mean_c.R
\name{phase_diff_mean_c}
\alias{phase_diff_mean_c}
\title{phase_diff_mean_c}
\usage{
phase_diff_mean_c(x1, x2, lowlim = 0, highlim = 360)
}
\arguments{
\item{x1, x2}{x coordinates to calculate phase difference between}

\item{lowlim, highlim}{maximum and minimum expected x values. Defaults apply to degrees in a circle}
}
\value{
Returns a vector of values (between -1 and 1) indicating the phase difference between two x coordinates
}
\description{
\code{phase_diff_mean_c} calcultes the phase difference between two x coordinates with a mean centred threshold.
Values more than 180 degrees ahead of the mean will be treated as lags, values less than 180 degrees behind mean
will be treated as leading
}
\examples{


x1 <- seq(0, 360, 360/4)
x2 <- 180

phase_diff(x1, x2)

}
