% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{word_aois}
\alias{word_aois}
\title{Words AOIs for eye_data_noisy and eye_data_clean.}
\format{A data frame with 17750 rows and 4 variables:
\describe{
\item{row}{line number}
\item{x1,x2,y1,y2}{word position}
}}
\usage{
word_aois
}
\description{
A dataset containing AOIs.
}
\keyword{datasets}
